backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
PyPDF2==3.0.1
werkzeug==3.0.1
pandas==2.2.1
matplotlib==3.8.3
python-jose==3.3.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0 